+++ pkg/rpc/server/server_test.go
@@ -1049,7 +1049,7 @@ func newBlock(t *testing.T, bc blockchainer.Blockchainer, index uint32, txs ...*
 	b := &block.Block{
 		Base: block.Base{
 			PrevHash:      hdr.Hash(),
-			Timestamp:     uint32(time.Now().UTC().Unix()) + hdr.Index,
+			Timestamp:     uint64(time.Now().UTC().Unix()) + uint64(hdr.Index),
 			Index:         hdr.Index + index,
 			ConsensusData: 1111,
 			NextConsensus: witness.ScriptHash(),
